Add erl modules showcasing longest path on DAG and child-process-over-port-communication
Longest path on a DAG:
* Node2 is the simple O(2^n) solution
* Node is the more efficient solution O(n^2), probably

Wanted to see how tough this problem was in Erlang. The simple case is very straightforward.
The solution where visited status is tracked is very gnarly due to the no side-effect
nature of Erlang.

Port communication to child process:
* complex.c is the child process "server code"
* port.c contains main routine, erl_comm.c reads from file descriptor 0 (unbuffered stdin)
  and writes to file desc 1 (unbuffered stdout.)

Wanted to see how tricky it is to do IPC over ports: it's not hard.