Add MathProg model to solve a basic vbucket connection assignment problem.
This is a subproblem of the general vbucket map generation problem:
basically this model creates a set of r sub-networks (one for each
replica) such that each possible node to node connection is
part of exactly one replica network.

The solution to this problem can then be fed into the vbucket
balancing problem - I believe without loss of generality.