Added prev param for active and replica vbuckets and for prev replication flow.
It's not clear that the prev flow is needed. Actually it looks like the
previous flow is not needed at all as the reductions in flow can be
equivalently captured by the moving around of active and replica
vbuckets.

It turns out that this version of the model has a fundamental problem
in that when adding a node the model produces a result that has active
and replica vbuckets moved to it that are the same vbuckets, which
clearly is broken.

This required a rethink moving to a model where source of vbuckets
are considered independently and can't mix. See the model that
introduces colors.