Fix issues with residual arcs and add new shortest path routine
Path.find_max_augmmenting_flow() wasn't respecting its color argument
which caused things to break badly.
Path.consume_flow() had a bug where we'd actually end up increasing the
flow at the source even as we thought we were decreasing it.
Added a new, supposedly more efficient shortest path routine that is
similar to the existing one but uses a min heap to track the unvisited
Nodes. It's Dijkstra-ish except that when it encounters a node whose
distance is reduced it adds it back to the unvisited list -- which I
think makes it safe for arcs with negative costs (which we have) but
messes up Dijkstra's nice worst case complexity. However, I think this
is near optimal for the data placement problem.

Issue is that there's a bug, which shows up for seven nodes with
six slaves. So that's next. :-)