Fixed computation of replica vbucket moves after solution
Previously if there were any paths through the network that did
things like "loop back" from a node in the prev_rvb layer to the
rvb layer the replica move computation was wrong. This is now
corrected: active and replica vbucket moves are computed in a
little more robust fashion.

Also fixed bug in arc where getting a flow for a color for
which there is no flow blew up. Also minor changes to formal
argument names in a couple of places.