Added support for min # of active and replica vbuckets per node
This is done with the addition of 2 extra layers: the avbp and
rvbp layers (meaning avb prime rvb prime.) 2 arcs are created
between each avb node and a corresponding avbp node: one
capacitated with the min # of active vbuckets on that node
with an attractive negative cost and a second arc that's free
with infinite capacity.

There are similar arcs between the rvbp and rvb layers.

Along with the upper bound on the replication links this should
cause the network to converge to be perfectly balanced with
one replica. This isn't the case though for 3 -> 4 node network
-- and this needs to be investigated.