Default the from address to the SMTP relay account user address
As it's easy to forget to explicitly provide the address and the
error message back from the SMTP server is a little bit inscrutable.
("Bad sender address syntax"). And it's a waste of time to figure it
out again.

Also improved the usage and changed the code to explicitly pass
the from address when calling sendmail.